D3D Branch nearly complete
- Fixed some issues with the texture support in the tile system
- Fixed some issues with GL api changes
- Fixed project settings and added d3d builds
- Depth Stencil now appears to be working
- Hid the flags variable in the texture system in OpenGL

+++ src/vcTerrain.cpp
@@ -58,10 +58,10 @@ void vcTerrain_BuildTerrain(vcTerrain *pTerrain, int16_t srid, const udDouble3 w
   int nodeCount = 0;
 
   // calculate view in quad tree space [0, 1]
-  float maxX = udMax(worldCorners[3].x, worldCorners[1].x);
-  float minX = udMin(worldCorners[0].x, worldCorners[2].x);
-  float maxY = udMax(worldCorners[0].y, worldCorners[1].y);
-  float minY = udMin(worldCorners[2].y, worldCorners[3].y);
+  float maxX = (float)udMax(worldCorners[3].x, worldCorners[1].x);
+  float minX = (float)udMin(worldCorners[0].x, worldCorners[2].x);
+  float maxY = (float)udMax(worldCorners[0].y, worldCorners[1].y);
+  float minY = (float)udMin(worldCorners[2].y, worldCorners[3].y);
   udDouble2 worldScale = udDouble2::create(maxX - minX, maxY - minY);
   udFloat2 viewPosMS = udFloat2::create((localViewPos.toVector2() - udDouble2::create(minX, minY)) / worldScale);
   udFloat2 viewSizeMS = udFloat2::create((float)localViewSize);
